requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.6
idna==3.10
importlib_resources==6.5.2
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
urllib3==2.3.0
zipp==3.21.0
